+++ CHANGELOG.md
@@ -34,6 +34,12 @@ console.log((await users.next()).page); // { total: 8268, offset: 60, limit: 20 
 console.log((await users.previous()).page); // { total: 8268, offset: 40, limit: 20 }
 ```
 
+### Changes
+
+- added `endTimestamp` to list of fieldnames that are parsed at Date
+- rqlBuilder now supports the `excludes` operator
+- rqlBuilder now accepts a regular string which will be checked using the parser function
+
 ## [v4.4.0]
 
 ### Changes